stable order inputs: find, sort, tar -T
After `cd`-ing to `$archdir`, use `find` and `sort` to generate a stable order
of inputs[0] (in the form of a newline-delimited list of paths) from which
`tar` reads via `-T-`.

This feature depends on `tar` support for the `-T` option as well as the
presence of the commands `find` and `sort`.

[0]:https://reproducible-builds.org/docs/stable-inputs/