Add cleanup script via `--cleanup file` argument (#201)
Use case example:
During execution a temporary Socker image is loaded and started, but
must be stopped and removed before returning to the controlling
terminal. Since this must happen in both exit cases (normal exit, and
signal capture handling) and since `trap` handling does not appear to work inside
`startup_script`, an alternative is needed.

Specifying `makeself ... --cleanup ./cleanup.sh ...` allows using an
included script to stop and remove the image in both situations.

The cleanup script is provided the same arguments that are given to
`startup_script` and is executed from the same working directory.