Feature: Allow extra options to compressor
This feature adds a new command line parameter, --comp-extra, which
functions similarly to --tar-extra, in that it allows a user to add
extra parameters to whichever compressor they have chosen to use.

The basis for this feature was to allow passing --no-name to gzip,
as part of a quest to make fully reproducible makeself archives.