Remove negative test
The negative test was written assuming gzip implementations would return
an error code when invalid command line options were used. This is not
always the case: BSD variants of gzip (e.g., FreeBSD, Apple) return
success even when an invalid option is specified.

Since this is out of the control of makeself, and behaviors will differ
widely among the various supported compressors and platforms, it doesn't
make sense to maintain this test. The same behavior is theoretically
possible with 'tar' implementations and the --tar-extra flag.